fix the profiles img and some buttons

+++ src/components/profiles/admin/TheClients.js
@@ -2,6 +2,7 @@ import React, { useEffect, useState } from "react";
 import axios from "axios";
 import cookie from "react-cookies";
 import { Table,Button } from "react-bootstrap";
+import Swal from "sweetalert2";
 function TheClients() {
   const [allUsers, setAllUsers] = useState([]);
   const token = cookie.load("token");
@@ -22,14 +23,56 @@ function TheClients() {
 
   // ::::::: handleDeleteUser::::::
   const handleDeleteUser = async (id) => {
+
+
+    const swalWithBootstrapButtons = Swal.mixin({
+      customClass: {
+        confirmButton: 'btn btn-success',
+        cancelButton: 'btn btn-danger'
+      },
+      buttonsStyling: false
+    })
+    
+    swalWithBootstrapButtons.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Yes, delete it!',
+      cancelButtonText: 'No, cancel!',
+      reverseButtons: true
+    }).then(async(result) => {
+      if (result.isConfirmed) {
+        let res = await axios.delete(`${Api}/deleteuser/${id}`, {
+          headers: {
+            authorization: `Bearer ${token}`,
+          },
+        });
+        setAllUsers(res.data);
+        swalWithBootstrapButtons.fire(
+          'Deleted!',
+          'Your file has been deleted.',
+          'success'
+        )
+      } else if (
+        /* Read more about handling dismissals below */
+        result.dismiss === Swal.DismissReason.cancel
+      ) {
+        swalWithBootstrapButtons.fire(
+          'Cancelled',
+          'Your imaginary file is safe :)',
+          'error'
+        )
+      }
+    })
     // /deleteuser/:id
 
-    let res = await axios.delete(`${Api}/deleteuser/${id}`, {
-      headers: {
-        authorization: `Bearer ${token}`,
-      },
-    });
-    setAllUsers(res.data);
+    // let res = await axios.delete(`${Api}/deleteuser/${id}`, {
+    //   headers: {
+    //     authorization: `Bearer ${token}`,
+    //   },
+    // });
+    // setAllUsers(res.data);
   };
 
   return (
